Adding demo automation script for local_dock_ssrA_peptide_against_sspB.
The original inputs and outputs are still included in directories named
'input_files.orig' and 'output_files.orig'. Each step in the command
file was tested on my machine and it works!

-------
Specifically, this revision fixes an error in prepack_flags and reduces the
number of strucutres built to 1 (from 3) for each range of motion to decrease
the runtime.