Adding public demo folder for FARFAR modeling guided by MOHCA-seq constraints, at demos/public/mohca_seq.
Includes input files and example output for helix pre-assembly, preparation of reference models, FARNA and FARFAR with MOHCA-seq constraints, clustering, and remodeling (example output not added for remodeling yet). All files use the ligand-binding region of a tandem glycine-binding riboswitch from F. nucleatum as an example system; the remodeling folder provides an example for adding a linker sequence and remodeling a linker region between the glycine-binding domains. The workflow in README.txt is described in the Methods in Enzymology chapter "Modeling complex RNA tertiary folds with Rosetta".