Updated so that testing server can test automatically. Instead of assuming the path to the Rosetta bin or having the user change the directory, the script will first check if the ROSETTA3 environment variable has been set. If so, use this as the EXE_DIR variable. Else, revert to the previous behavior.
We have done a similar thing for the Rosetta/tools/ directory, which will look first for a $TOOLS environment variable.

We also needed to make sure linuxgccrelease and macosgccrelease were not hardcoded in, so another command line variable that is now available is "exe_extension=". The argument parser looks for a variable that begins with "exe_extension=" and will use this extension if provided. If the user does not use this argument, the previous behavior is assumed.